# docs/swagger.yaml
basePath: /
definitions:
...
  models.Pet:
    description: 宠物信息模型
    properties:
      age:
        description: Age is the age of the pet in months
        example: 6
        type: integer
      ageDisplay:
        description: AgeDisplay is a human-readable age string (e.g., "2个月", "1岁")
        example: 6个月
        type: string
      breed:
        description: Breed is the specific breed of the pet
        example: 金毛寻回犬
        type: string
      createdAt:
        description: CreatedAt is when the pet was added
        example: "2024-01-10"
        type: string
      description:
        description: Description is a detailed description of the pet
        example: 活泼可爱的金毛宝宝
        type: string
      healthStatus:
        description: HealthStatus describes the health condition of the pet
        example: 健康状况良好
        type: string
      id:
        description: ID is the unique identifier for the pet
        example: 1
        type: integer
      name:
        description: Name is the name of the pet
        example: 旺财
        type: string
      photoUrls:
        description: PhotoUrls contains a list of photo URLs for the pet
        items:
          type: string
        type: array
      price:
        description: Price is the price of the pet in CNY
        example: 3500
        type: number
      status:
        description: Status indicates the availability of the pet (e.g., "available",
          "pending", "sold")
        example: available
        type: string
      type:
        description: Type is the species or breed of the pet (e.g., "Dog", "Cat")
        example: 狗狗
        type: string
      vaccinationRecords:
        description: VaccinationRecords contains vaccination information
        items:
          $ref: '#/definitions/models.VaccinationRecord'
        type: array
    type: object
...
  models.VaccinationRecord:
    properties:
      completed:
        type: boolean
      date:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PetShop 宠物商店 API 文档 - 提供宠物管理、购物车、订单管理、用户管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PetShop API
  version: "1.0"
paths:
...
  /api/pet:
    delete:
      consumes:
      - application/json
      description: 根据ID删除宠物
      parameters:
      - description: 宠物ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 已删除的宠物信息
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 宠物未找到
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除宠物
      tags:
      - 宠物管理
    get:
      consumes:
      - application/json
      description: 根据宠物ID获取详细信息，支持缓存
      parameters:
      - description: 宠物ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 宠物详情
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 宠物未找到
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取宠物详情
      tags:
      - 宠物管理
    put:
      consumes:
      - application/json
      description: 更新宠物的详细信息，支持部分字段更新
      parameters:
      - description: 宠物信息
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.Pet'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的宠物信息
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 宠物未找到
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新宠物
      tags:
      - 宠物管理
...
  /api/pet/search:
    get:
      consumes:
      - application/json
      description: 根据宠物名称搜索，支持分页
      parameters:
      - description: 宠物名称关键词
        in: query
        name: name
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 搜索宠物
      tags:
      - 宠物管理
  /api/pets:
    get:
      consumes:
      - application/json
      description: 获取分页的宠物列表，支持按类型筛选
      parameters:
      - description: 宠物类型筛选
        in: query
        name: type
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页的宠物列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取宠物列表
      tags:
      - 宠物管理
...
  /api/v1/pets:
    get:
      consumes:
      - application/json
      description: 根据类型、状态、价格范围、搜索关键词等条件筛选宠物
      parameters:
      - description: 宠物类型
        in: query
        name: type
        type: string
      - description: 宠物状态 (available, pending, sold)
        in: query
        name: status
        type: string
      - description: 最低价格
        in: query
        name: minPrice
        type: number
      - description: 最高价格
        in: query
        name: maxPrice
        type: number
      - description: 搜索关键词（名称或品种）
        in: query
        name: search
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 筛选结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 筛选宠物
      tags:
      - 宠物管理
...
securityDefinitions:
  APITokenAuth:
    description: 使用 API Token 进行认证，用于 Open API 访问
    in: header
    name: X-API-Token
    type: apiKey
  BearerAuth:
    description: 使用 JWT Token 进行认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"